--- src/cmps/Footer.tsx
@@ -1,21 +1,0 @@
-function Footer() {
-
-    return (
-        <div className="Footer"
-            style={{
-                backgroundColor: "#DEF5E5",
-                height: "10vh", 
-                padding: "10px",
-                alignItems: "center",
-                display: "inline-flexbox",
-                flexDirection: "column",
-                justifyContent: "center"
-            }}
-        >
-            <h1>Footer</h1>
-        </div >
-    );
-
-}
-
-export default Footer;
