--- src/cmps/Header.tsx
@@ -1,22 +1,0 @@
-function Header() {
-
-
-    return (
-      <div className="Header"
-        style={{
-          backgroundColor: "#8EC3B0",
-          height: "10vh",
-          padding: "10px",
-          display: "inline-flexbox",
-          flexDirection: "column",
-          justifyContent: "center"
-        }}
-      >
-          <h1>Header</h1>
-      </div >
-    );
-  
-  }
-  
-  export default Header;
-
