+++ src/components/Footer.tsx
@@ -1,0 +1,17 @@
+import { Box } from '@chakra-ui/react';
+
+function Footer() {
+
+    return (
+        <Box className="Footer"
+            bg="#DEF5E5"
+            h="10vh"
+            p={2.5}
+        >
+            <h1>Footer</h1>
+        </Box >
+    );
+
+}
+
+export default Footer;
