+++ src/components/Header.tsx
@@ -1,0 +1,18 @@
+import { Box } from '@chakra-ui/react';
+
+
+function Header() {
+    return (
+      <Box className="Header"
+      bg="#8EC3B0"
+      h="10vh"
+      p={2.5}
+      >
+          <h1>Header</h1>
+      </Box >
+    );
+  
+  }
+  
+  export default Header;
+
