+++ src/index.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { ChakraProvider  } from "@chakra-ui/react"
 import './index.css';
-import Header from './cmps/Header';
+import Header from './components/Header';
 import App from './App';
-import Footer from './cmps/Footer';
+import Footer from './components/Footer';
 import reportWebVitals from './reportWebVitals';
 
 const root = ReactDOM.createRoot(
